% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{.materialize_ref}
\alias{.materialize_ref}
\title{Materialize ref binding into a resolution descriptor}
\usage{
.materialize_ref(binding, row, resources = NULL)
}
\description{
Materialize ref binding into a resolution descriptor
}
\keyword{internal}
