% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{.resolve_via_backend}
\alias{.resolve_via_backend}
\title{Dispatch to backend adapter for resolution}
\usage{
.resolve_via_backend(ref_info, logical_schema, root = NULL, as_array = TRUE)
}
\description{
Dispatch to backend adapter for resolution
}
\keyword{internal}
