% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{.validate_resolved}
\alias{.validate_resolved}
\title{Validate resolved value against logical schema}
\usage{
.validate_resolved(value, logical_schema)
}
\description{
Validate resolved value against logical schema
}
\keyword{internal}
