% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{encoding_applicable}
\alias{encoding_applicable}
\title{Check if an encoding is applicable for a given row}
\usage{
encoding_applicable(encoding, row)
}
\description{
Check if an encoding is applicable for a given row
}
\keyword{internal}
