% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurotabs-package.R
\docType{package}
\name{neurotabs-package}
\alias{neurotabs}
\alias{neurotabs-package}
\title{neurotabs: Neuro Feature Tables - Storage-Independent Neuroimaging Datasets}
\description{
Reference implementation of the NFTab specification for row-oriented neuroimaging datasets. Read, write, validate, and query collections of neuroimaging observations with design metadata and feature values that resolve to logical arrays (volumes, ROI vectors, surfaces) independent of physical storage backend.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/neurotabs/}
  \item \url{https://github.com/bbuchsbaum/neurotabs}
  \item Report bugs at \url{https://github.com/bbuchsbaum/neurotabs/issues}
}

}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
