% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{nf_axis_info}
\alias{nf_axis_info}
\title{Axis metadata for a feature axis}
\usage{
nf_axis_info(x, feature, axis)
}
\arguments{
\item{x}{An nftab object.}

\item{feature}{Feature name.}

\item{axis}{Axis name from the feature's logical schema.}
}
\value{
An \code{nf_axis_domain} object.
}
\description{
Axis metadata for a feature axis
}
