% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{nf_axis_labels}
\alias{nf_axis_labels}
\title{Read axis labels for a feature axis}
\usage{
nf_axis_labels(x, feature, axis)
}
\arguments{
\item{x}{An nftab object.}

\item{feature}{Feature name.}

\item{axis}{Axis name from the feature's logical schema.}
}
\value{
A data.frame with at least \code{index} and \code{label}.
}
\description{
Read axis labels for a feature axis
}
