% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{nf_col}
\alias{nf_col}
\title{Create a column reference (ValueSource)}
\usage{
nf_col(column_name)
}
\arguments{
\item{column_name}{Name of the observation table column.}
}
\value{
A list with class \code{nf_column_ref}.
}
\description{
Create a column reference (ValueSource)
}
