% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{nf_compatible}
\alias{nf_compatible}
\title{Check strict concatenation compatibility}
\usage{
nf_compatible(a, b)
}
\arguments{
\item{a}{An \link{nftab} object.}

\item{b}{An \link{nftab} object.}
}
\value{
A list with \code{compatible} (logical) and \code{reasons} (character vector
of incompatibilities). Returned invisibly.
}
\description{
Check strict concatenation compatibility
}
