% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{nf_concat}
\alias{nf_concat}
\title{Concatenate NFTab datasets (strict row-wise)}
\usage{
nf_concat(..., provenance_col = "source_dataset")
}
\arguments{
\item{...}{\link{nftab} objects to concatenate.}

\item{provenance_col}{Name of provenance column to add. Default
\code{"source_dataset"}. Set to \code{NULL} to skip.}
}
\value{
A new \link{nftab} object with rows from all inputs.
}
\description{
Concatenate NFTab datasets (strict row-wise)
}
