% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{nf_design}
\alias{nf_design}
\title{Design columns (observation table as data.frame)}
\usage{
nf_design(x)
}
\arguments{
\item{x}{An nftab object.}
}
\value{
A data.frame.
}
\description{
Design columns (observation table as data.frame)
}
