% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{nf_extension}
\alias{nf_extension}
\title{Retrieve one manifest extension}
\usage{
nf_extension(x, key)
}
\arguments{
\item{x}{An nftab object.}

\item{key}{Extension key, including the \verb{x-} prefix.}
}
\value{
The extension value.
}
\description{
Retrieve one manifest extension
}
