% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{nf_extensions}
\alias{nf_extensions}
\title{Manifest extension entries}
\usage{
nf_extensions(x)
}
\arguments{
\item{x}{An nftab object.}
}
\value{
A named list of manifest extensions, or \code{NULL}.
}
\description{
Manifest extension entries
}
