% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{nf_feature}
\alias{nf_feature}
\title{Declare a feature schema}
\usage{
nf_feature(logical, encodings, nullable = FALSE, description = NULL)
}
\arguments{
\item{logical}{An \link{nf_logical_schema} object.}

\item{encodings}{A list of encoding objects created by \code{\link[=nf_ref_encoding]{nf_ref_encoding()}} or
\code{\link[=nf_columns_encoding]{nf_columns_encoding()}}.}

\item{nullable}{Whether missing values are permitted. Default \code{FALSE}.}

\item{description}{Optional description.}
}
\value{
An \code{nf_feature} object.
}
\description{
Declare a feature schema
}
