% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{nf_feature_names}
\alias{nf_feature_names}
\title{Feature names}
\usage{
nf_feature_names(x)
}
\arguments{
\item{x}{An nftab object.}
}
\value{
Character vector of feature names.
}
\description{
Feature names
}
