% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{nf_feature_schema}
\alias{nf_feature_schema}
\title{Logical schema for a feature}
\usage{
nf_feature_schema(x, feature)
}
\arguments{
\item{x}{An nftab object.}

\item{feature}{Feature name.}
}
\value{
An \code{nf_logical_schema} object.
}
\description{
Logical schema for a feature
}
