% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{nf_filter}
\alias{nf_filter}
\title{Filter observations by design predicates}
\usage{
nf_filter(x, ...)
}
\arguments{
\item{x}{An \link{nftab} object.}

\item{...}{Filter expressions evaluated in the context of the observation
table. Observation column names take precedence over variables in the
calling environment. Multiple expressions are combined with AND.}
}
\value{
A filtered \link{nftab} object.
}
\description{
Filter observations by design predicates
}
