% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{nf_read}
\alias{nf_read}
\title{Read an NFTab dataset from a manifest file}
\usage{
nf_read(path, validate_schema = TRUE)
}
\arguments{
\item{path}{Path to \code{nftab.yaml} or \code{nftab.json}.}

\item{validate_schema}{Whether to validate the manifest against the bundled
JSON Schema. Default \code{TRUE}.}
}
\value{
An \link{nftab} object.
}
\description{
Read an NFTab dataset from a manifest file
}
