% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{nf_support_info}
\alias{nf_support_info}
\title{Retrieve one manifest support}
\usage{
nf_support_info(x, ref)
}
\arguments{
\item{x}{An nftab object.}

\item{ref}{Manifest-local support reference.}
}
\value{
An \code{nf_support_schema} object.
}
\description{
Retrieve one manifest support
}
