% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{nf_supports}
\alias{nf_supports}
\title{Manifest supports}
\usage{
nf_supports(x)
}
\arguments{
\item{x}{An nftab object.}
}
\value{
A named list of support descriptors, or \code{NULL}.
}
\description{
Manifest supports
}
