% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{nf_ungroup}
\alias{nf_ungroup}
\title{Drop grouping metadata}
\usage{
nf_ungroup(x)
}
\arguments{
\item{x}{A \code{grouped_nftab} object.}
}
\value{
The underlying \link{nftab} object.
}
\description{
Drop grouping metadata
}
