% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{nf_write}
\alias{nf_write}
\title{Write an NFTab dataset to disk (table-package profile)}
\usage{
nf_write(x, path, manifest_name = "nftab.yaml")
}
\arguments{
\item{x}{An \link{nftab} object.}

\item{path}{Directory to write the dataset into. Created if it doesn't exist.}

\item{manifest_name}{Manifest filename. Default \code{"nftab.yaml"}.}
}
\value{
Invisibly, the path written to.
}
\description{
Write an NFTab dataset to disk (table-package profile)
}
