% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{resolve_value_source}
\alias{resolve_value_source}
\title{Resolve a ValueSource for a given row}
\usage{
resolve_value_source(source, row)
}
\description{
Resolve a ValueSource for a given row
}
\keyword{internal}
