% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nftab-class.R
\name{[.nftab}
\alias{[.nftab}
\title{Subset an nftab by row indices}
\usage{
\method{[}{nftab}(x, i, ...)
}
\arguments{
\item{x}{An nftab object.}

\item{i}{Row indices.}

\item{...}{Ignored.}
}
\description{
Subset an nftab by row indices
}
